mare

+++ inc/Map.class.hpp
@@ -38,7 +38,7 @@ class Map
     	void		set_water(int x, int y, float value);
 		int			get_hauteur_max(void) const ;
     	void		apply_gravity(void);
-    	void		elevate_rect(int x0, int y0, int x1, int y1, int value);
+    	void		elevate_rect(int x0, int y0, int x1, int y1, float value);
         void		draw_cone(int start_x, int start_y, int radius, int height, bool reverse = false);
 	    std::string	to_string(void) const;
         int			resistance(int x, int y);
